+++ R/rpanet_internal.R
@@ -26,11 +26,13 @@ NULL
 #' non-linear (\code{rpanet_general}) preference functions
 #'
 #' @param nstep Number of steps when generating a network.
-#' @param initial.network A \code{wdnet} object or a list representing the seed
-#'   network. By default, \code{initial.network} has one edge from node 1 to
-#'   node 2 with weight 1. It may contain the following components: a two column
+#' @param initial.network A \code{wdnet} object or a list that represents the
+#'   initial network. By default, \code{initial.network} has one directed edge from node 1
+#'   to node 2 with weight 1. It may have the following components: a two-column
 #'   matrix \code{edgelist} representing the edges; a vector \code{edgeweight}
-#'   representing the weight of edges; an integer vector \code{nodegroup}
+#'   representing the weight of edges; a logical argument \code{directed} indicating
+#'   whether the initial network is directed;
+#'   an integer vector \code{nodegroup}
 #'   representing the group of nodes. \code{nodegroup} is defined for directed
 #'   networks, if \code{NULL}, all nodes from the seed network are considered
 #'   from group 1.
@@ -39,8 +41,6 @@ NULL
 #'   each step, a new edge of weight 1 is added from a new node \code{A} to an
 #'   existing node \code{B} (\code{alpha} scenario), where $\code{B} is chosen
 #'   with probability proportional to its in-strength + 1.
-#' @param directed Logical, whether to generate directed networks. If
-#'   \code{FALSE}, the edge directions are ignored.
 #' @param m Integer vector, number of new edges in each step.
 #' @param sum_m Integer, summation of \code{m}.
 #' @param w Vector, weight of new edges.
@@ -72,7 +72,7 @@ NULL
 #' @keywords internal
 #' 
 rpanet_general <- function(
-    nstep, initial.network, control, directed,
+    nstep, initial.network, control,
     m, sum_m, w,
     nnode, nedge, method, sample.recip) {
   edgeweight <- c(initial.network$edge.attr$weight, w)
@@ -88,7 +88,7 @@ rpanet_general <- function(
     control$preference$ftype == "default",
     1, 2
   )
-  if (directed) {
+  if (initial.network$directed) {
     outs <- double(node_vec_length)
     ins <- double(node_vec_length)
     outs[1:nnode] <- initial.network$node.attr$outs
@@ -187,7 +187,7 @@ rpanet_general <- function(
       ),
       "newedge" = ret_c$m,
       "control" = control,
-      "directed" = directed,
+      "directed" = initial.network$directed,
       "edge.attr" = data.frame(
         "weight" = edgeweight[1:nedge],
         "scenario" = ret_c$scenario[1:nedge]
@@ -196,7 +196,7 @@ rpanet_general <- function(
     class = "wdnet"
   )
   ret$weighted <- any(ret$edge.attr$weight != 1)
-  if (directed) {
+  if (initial.network$directed) {
     ret$node.attr <- data.frame(
       "outs" = ret_c$outs[1:nnode],
       "ins" = ret_c$ins[1:nnode],
@@ -224,7 +224,7 @@ rpanet_general <- function(
 #' @keywords internal
 #'
 rpanet_simple <- function(
-    nstep, initial.network, control, directed,
+    nstep, initial.network, control,
     m, sum_m, w, ex_node, ex_edge, method) {
   delta <- control$preference$params[2]
   delta_out <- control$preference$sparams[5]
@@ -240,7 +240,7 @@ rpanet_simple <- function(
       control$scenario$rho
     )
   )
-  if (!directed) {
+  if (!initial.network$directed) {
     delta_out <- delta_in <- delta / 2
   }
   if (method == "bag") {
@@ -251,7 +251,7 @@ rpanet_simple <- function(
       scenario,
       ex_node, ex_edge,
       delta_out, delta_in,
-      directed
+      initial.network$directed
     )
     snode <- ret$snode
     tnode <- ret$tnode
@@ -300,7 +300,7 @@ rpanet_simple <- function(
       left.open = TRUE
     )
     end_edge <- findInterval(rand_in[temp_in], weight_intv, left.open = TRUE)
-    if (directed) {
+    if (initial.network$directed) {
       snode <- find_node_cpp(snode, start_edge)
       tnode <- find_node_cpp(tnode, end_edge)
     } else {
@@ -317,11 +317,11 @@ rpanet_simple <- function(
     edgeweight = edgeweight,
     newedge = m,
     control = control,
-    directed = directed,
+    directed = initial.network$directed,
     weighted = any(edgeweight != 1)
   )
   ret$edge.attr$scenario <- c(rep(0, ex_edge), scenario)
-  if (directed) {
+  if (initial.network$directed) {
     ret$node.attr$spref <- ret$node.attr$outs +
       control$preference$sparams[5]
     ret$node.attr$tpref <- ret$node.attr$ins +